flycatch_auth/pyproject.toml
[tool.poetry]
name = "flycatch_auth"
version = "0.1.0"
description = "A Python authentication package supporting Flask and FastAPI"
authors = ["flycatch"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yourusername/flycatch_auth"
packages = [{ include = "flycatch_auth" }]
keywords = ["authentication", "jwt", "flask", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]

[tool.poetry.dependencies]
python = ">=3.11"
flask = "^3.1.0"
fastapi = "^0.115.8"
pyjwt = "^2.10.1"
pydantic = "^2.10.6"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
